Version 9.0.2102.0501
https://www.gurux.fi/node/17806

+++ Development/GXDLMSXmlClient.cs
@@ -41,6 +41,7 @@ using Gurux.DLMS.Enums;
 using System.Xml;
 using System.IO;
 using System.ComponentModel;
+using Gurux.DLMS.Objects.Enums;
 
 namespace Gurux.DLMS
 {
@@ -631,7 +632,7 @@ namespace Gurux.DLMS
                 GXByteBuffer reply;
                 if (Settings.InterfaceType == InterfaceType.WRAPPER)
                 {
-                    if (Ciphering.Security != (byte)Security.None)
+                    if (Ciphering.Security != Security.None)
                     {
                         GXDLMSLNParameters p = new GXDLMSLNParameters(this, Settings, 0, pdu.Command, 0x0, null, null, 0xff, Command.None);
                         reply = new GXByteBuffer(GXDLMS.Cipher0(p, pdu.Data));
@@ -643,7 +644,7 @@ namespace Gurux.DLMS
                 }
                 else
                 {
-                    if (Ciphering.Security != (byte)Security.None)
+                    if (Ciphering.Security != Security.None)
                     {
                         GXDLMSLNParameters p = new GXDLMSLNParameters(this, Settings, 0, pdu.Command, 0x0, null, null, 0xff, Command.None);
                         byte[] tmp = GXDLMS.Cipher0(p, pdu.Data);
